backend/requirements.txt
Django>=5.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf-yasg
gunicorn
whitenoise
django-extensions
psycopg2-binary